+++ lib/puppet/type_guts/attribute_definitions.rb
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 module Puppet
   class Type
     newmetaparam(:noop) do
