+++ lib/puppet/type_guts/automatic_relationships.rb
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 module Puppet
   class Type
     # Adds a block producing a single name (or list of names) of the given
