+++ lib/puppet/type_guts/comparable.rb
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 module Puppet
   class Type
     include Comparable
