+++ lib/puppet/type_guts/device_applicability.rb
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 
 module Puppet
   class Type
